Fix RSS feed
My last commit added a default title for title-less posts that was a
pretty version of the publish date. Unfortunately this added a title
to title-less posts in my RSS feed, which made for an ugly timeline and
also messed with how micro.blog formats its timeline (posts with a tile
just show the title and a permalink, vs showing the post body). The
unfortunate side-effect of this is that I had to make title-less post
pages have a site-level title (i.e. "Pete Schaffner").